refactor(tokenizer): organiza funcionalidades de tokenização e aplica norminete
