+++ Problemas.md
@@ -1,0 +1,4 @@
+# Identificação de problemas
+
+- [ ] `cat Makefile |>out` fica travado parecendo que está esperando algo 
+	- O que é estranho pois quando executado `cat Makefile |>out grep ".c"` ele funciona normalmente
