+++ macros.h
@@ -6,7 +6,7 @@
 /*   By: jlacerda <jlacerda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 19:15:26 by peda-cos          #+#    #+#             */
-/*   Updated: 2025/03/20 22:48:48 by jlacerda         ###   ########.fr       */
+/*   Updated: 2025/03/22 15:49:41 by jlacerda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
